Rename HealthPaneItem to prep for D2772095 [Health Gadget 1/5]
Reviewed By: ssorallen

Differential Revision: D2772124

fb-gh-sync-id: 883ba2bfc1612b7580d92277ad8c6a1636268c7d
